Move util.IgnoreError to errorutil.Ignore  -Part of Mesos-DNS#223, to be explicit about ignoring errors
